// src/types.ts
export type UserRole = 'PATIENT' | 'DOCTOR';

export interface KeyPair {
  publicKey: string;
  privateKey: string;
}

export interface UserProfile {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  licenceNumber?: string; // For Doctors
  specialty?: string; // For Doctors
  publicKey: string;
  privateKey?: string; // Stored securely on client-side for patients (simulation)
  createdAt: string;
  passcode?: string; // Secure passcode/password configured on account creation
}

export interface MedicalRecord {
  id: string;
  patientId: string;
  title: string;
  category: 'Prescription' | 'Lab Report' | 'Diagnosis' | 'Immunization';
  doctorName: string;
  date: string;
  details: string; // Plaintext (before encryption)
  encryptedContent: string; // Simulated cipher text
  medicines: string[]; // List of drugs for AI scanner
  hash: string; // SHA-256 integrity simulation
  photoUrl?: string; // Optional patient record preview attachment photo
}

export interface AccessRequest {
  id: string;
  patientId: string;
  patientName: string;
  doctorId: string;
  doctorName: string;
  doctorSpecialty: string;
  requestedPurpose: string;
  status: 'PENDING' | 'APPROVED' | 'REJECTED' | 'REVOKED';
  requestedAt: string;
  decidedAt?: string;
  digitalSignature?: string; // Digital signature generated by patient's private key
}

export interface AuditLog {
  id: string;
  timestamp: string;
  action: string;
  actorName: string;
  actorRole: UserRole;
  status: 'SUCCESS' | 'WARNING' | 'ALERT';
  hash: string; // Simulation of immutable ledger block hash
  details: string;
}

export interface DrugConflictWarning {
  severity: 'CRITICAL' | 'MODERATE' | 'NONE';
  description: string;
  resolution: string;
}